BACKPORT: arm64: Add support for SB barrier and patch in over DSB; ISB sequences
We currently use a DSB; ISB sequence to inhibit speculation in set_fs().
Whilst this works for current CPUs, future CPUs may implement a new SB
barrier instruction which acts as an architected speculation barrier.

On CPUs that support it, patch in an SB; NOP sequence over the DSB; ISB
sequence and advertise the presence of the new instruction to userspace.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Change-Id: I28e2640aff2734576f64c508cef83ad411e8d3ab
Signed-off-by: Blaster4385 <venkateshchaturvedi12@gmail.com>
Signed-off-by: Pajdek <pajdek@gmail.com>