kernel: Boost whenever a zygote-forked process becomes a top app
Boost the CPU & CPUBW to the max for 500 ms whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[0ctobot: Adapt for 4.19]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: I512dde126e475945a9e784952c5ae689453c41c2
Signed-off-by: Pajdek <pajdek@gmail.com>