BACKPORT: sched: Rework CPU hotplug task selection
(cherry picked from commit 10e7071b2f491)

The CPU hotplug task selection is the only place where we used
put_prev_task() on a task that is not current. While looking at that,
it occured to me that we can simplify all that by by using a custom
pick loop.

Since we don't need to put current, we can do away with the fake task
too.

BUG=b:152605392
TEST=run power_VideoCall test

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Cc: Aaron Lu <aaron.lwe@gmail.com>
Cc: Valentin Schneider <valentin.schneider@arm.com>
Cc: mingo@kernel.org
Cc: Phil Auld <pauld@redhat.com>
Cc: Julien Desfossez <jdesfossez@digitalocean.com>
Cc: Nishanth Aravamudan <naravamudan@digitalocean.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: I75315b91d18d47464519cbccc89c070af08773c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2186285
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: Blaster4385 <venkateshchaturvedi12@gmail.com>
Signed-off-by: Pajdek <pajdek@gmail.com>