ion: Mark workqueues freeing buffers asynchronously as CPU intensive
When exiting the camera, there's a period of intense lag caused by all
of the buffer-free workers consuming all CPUs at once for a few seconds.
This isn't very good, and freeing the buffers isn't super time critical,
so we can lower the burden of the workers by marking the per-heap
workqueues as CPU intensive, which offloads the burden of balancing the
workers onto the scheduler.

Also, mark these workqueues with WQ_MEM_RECLAIM so forward progress is
guaranteed via a rescuer thread, since these are used to free memory.
The unnecessary WQ_UNBOUND_MAX_ACTIVE is removed as well, since it's
only used for increasing the active worker count on large-CPU systems.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I95282fed62d3bb70a8d8d17092aa95a8820ec0a7
Signed-off-by: Pajdek <pajdek@gmail.com>