power: oplus_vooc: Condtionally skip reading header/footer from vooc fw
OOS seems to be loading firmware while skipping the 80 byte header due to its userspace modifications, Custom ROMs do not skip that 80 byte header which results in vooc fw not being loaded. Fix this by checking if the firmware_data still has the 80 byte header after reading it from sysfs firmware file so that we can skip the header/footer if not already done.

Co-authored-by: Abhay Singh Gill <abhaygill017@gmail.com>
Suggested-by: jabashque <jabashque@gmail.com>
Signed-off-by: Abhay Singh Gill <abhaygill017@gmail.com>
Signed-off-by: Pajdek <pajdek@gmail.com>